fix: remove ShowPins.js and refactoring Map.js
