+++ src/pages/Pinboard.jsx
@@ -4,11 +4,8 @@ import styled from "styled-components";
 import { THEME } from "../constants/colors";
 import { useState } from "react";
 import React from "react";
-import { useEffect } from "react";
-import { showPins } from "../components/atoms/showpins";
 import Map from "../components/atoms/Map";
-
-const { kakao } = window;
+import pinData from "../pinData";
 
 const FormWrapper = styled.form`
     height: 100%;
@@ -21,14 +18,6 @@ const FormWrapper = styled.form`
 // 나중에 여기서 타입을 선택하든, 검색하든 백엔드에서 핀 정보 받아와서 그 핀을 showPins로 넘겨줘야 할 듯
 
 export default function Pinboard() {
-    //지도 및 핀 표시
-    useEffect(() => {
-        showPins();
-    }, []);
-
-    //목데이터를 받아서
-
-    //가공해서 ()
     const navigate = useNavigate();
 
     const [search, setSearch] = useState("");
@@ -40,7 +29,7 @@ export default function Pinboard() {
         <>
             <Layout title="Main Page">
                 {/* 지도 틀 */}
-                <Map></Map>
+                <Map pinData={pinData}></Map>
             </Layout>
         </>
     );
